pyproject.toml
[project]
name = "npu-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.2.6",
    "matplotlib>=3.8.0",
    "tabulate>=0.9.0",
    "pre-commit>=4.5.1",
    "torch>=2.10.0",
    "iree-base-compiler",
    "iree-base-runtime",
]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "ty>=0.0.29",
]
